+++ KoS.Apps.SharePoint.SmartCAML/KoS.Apps.SharePoint.SmartCAML.Editor/AttachedProperties/PasswordBoxAssistant.cs
@@ -5,17 +5,80 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.AttachedProperties
 {
     public static class PasswordBoxAssistant
     {
-        public static readonly DependencyProperty BoundPassword =
-            DependencyProperty.RegisterAttached("BoundPassword", typeof (string), typeof (PasswordBoxAssistant),
-                new PropertyMetadata(string.Empty, OnBoundPasswordChanged));
+        public static readonly DependencyProperty BoundPassword = DependencyProperty.RegisterAttached(
+                "BoundPassword", 
+                typeof (string), 
+                typeof (PasswordBoxAssistant),
+                new PropertyMetadata(string.Empty, OnBoundPasswordChanged)
+            );
 
         public static readonly DependencyProperty BindPassword = DependencyProperty.RegisterAttached(
-            "BindPassword", typeof (bool), typeof (PasswordBoxAssistant),
-            new PropertyMetadata(false, OnBindPasswordChanged));
+                "BindPassword", 
+                typeof (bool), 
+                typeof (PasswordBoxAssistant),
+                new PropertyMetadata(false, OnBindPasswordChanged)
+            );
 
-        private static readonly DependencyProperty UpdatingPassword =
-            DependencyProperty.RegisterAttached("UpdatingPassword", typeof (bool), typeof (PasswordBoxAssistant),
-                new PropertyMetadata(false));
+        public static readonly DependencyProperty BindPlaceholder = DependencyProperty.RegisterAttached(
+                "BindPlaceholder",
+                typeof(UIElement),
+                typeof(PasswordBoxAssistant),
+                new PropertyMetadata(null, OnBindPlaceholderChanged)
+            );
+
+        private static readonly DependencyProperty UpdatingPassword = DependencyProperty.RegisterAttached(
+                "UpdatingPassword", 
+                typeof (bool), 
+                typeof (PasswordBoxAssistant),
+                new PropertyMetadata(false)
+            );
+
+        public static UIElement GetBindPlaceholder(DependencyObject dp)
+        {
+            return (UIElement)dp.GetValue(BindPlaceholder);
+        }
+
+        public static void SetBindPlaceholder(DependencyObject dp, UIElement value)
+        {
+            dp.SetValue(BindPlaceholder, value);
+        }
+
+        private static void OnBindPlaceholderChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
+        {
+            var box = (PasswordBox)d;
+
+            if (e.OldValue as UIElement != null) {
+                box.PasswordChanged -= ShowHidePlaceholderOnPasswordChanged;
+                box.IsVisibleChanged -= Box_IsVisibleChanged;
+            }
+
+            if (e.NewValue as UIElement != null) {
+                box.PasswordChanged += ShowHidePlaceholderOnPasswordChanged;
+                box.IsVisibleChanged += Box_IsVisibleChanged;
+            }
+        }
+
+        private static void Box_IsVisibleChanged(object sender, DependencyPropertyChangedEventArgs e)
+        {
+            var box = (PasswordBox)sender;
+            ShowHidePlaceholder(box);
+        }
+
+        private static void ShowHidePlaceholderOnPasswordChanged(object sender, RoutedEventArgs e)
+        {
+            var box = (PasswordBox)sender;
+            ShowHidePlaceholder(box);
+        }
+
+        private static void ShowHidePlaceholder(PasswordBox box)
+        {
+            var placeholder = GetBindPlaceholder(box);
+
+            if (box.Visibility == Visibility.Visible && box.Password == string.Empty)
+                placeholder.Visibility = Visibility.Visible;
+            else
+                placeholder.Visibility = Visibility.Collapsed;
+        }
 
         private static void OnBoundPasswordChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
         {