+++ KoS.Apps.SharePoint.SmartCAML/KoS.Apps.SharePoint.SmartCAML.Editor/Dialogs/AboutWindow.xaml.cs
@@ -107,19 +107,6 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.Dialogs
             );
         }
 
-        private void DonateButton_OnClick(object sender, RoutedEventArgs e)
-        {
-            Telemetry.Instance.Native.TrackEvent("About.Donate");
-            try
-            {
-                Process.Start("https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=HTEBZ3Y37F2ZL");
-            }
-            catch (Exception ex)
-            {
-                ExceptionHandler.Handle(ex);
-            }
-        }
-
         private void UpdateStatusSuccess(string message, bool? installButtonVisible = null)
         {
             ucUpdateMessage.Foreground = Brushes.ForestGreen;