+++ KoS.Apps.SharePoint.SmartCAML/KoS.Apps.SharePoint.SmartCAML.Editor/UserControls/ConnectWindow.xaml.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.IO;
-using System.Net;
 using KoS.Apps.SharePoint.SmartCAML.SharePointProvider;
 using System.Windows;
 using System.Windows.Controls;
@@ -10,24 +8,21 @@ using KoS.Apps.SharePoint.SmartCAML.Editor.Model;
 using KoS.Apps.SharePoint.SmartCAML.Editor.Utils;
 using KoS.Apps.SharePoint.SmartCAML.Model;
 
-namespace KoS.Apps.SharePoint.SmartCAML.Editor.Dialogs
+namespace KoS.Apps.SharePoint.SmartCAML.Editor.UserControls
 {
     /// <summary>
     /// Interaction logic for ConnectWindow.xaml
     /// </summary>
-    public partial class ConnectWindow : Window
+    public partial class ConnectWindow : UserControl
     {
         public ISharePointProvider Client { get; private set; }
-        public ConnectWindowModel Model => (ConnectWindowModel) this.DataContext;
+        public ConnectWindowModel Model => (ConnectWindowModel)this.DataContext;
+        public event Action<ConnectWindow, ISharePointProvider> DialogResult;
 
         public ConnectWindow()
         {
             Telemetry.Instance.Native.TrackPageView("Connect");
             InitializeComponent();
-
-#if DEBUG
-            ucApiFake.Visibility = Visibility.Visible;
-#endif
 
             this.Width = Config.ConnectWindowWidth;
             this.DataContext = new ConnectWindowModel();
@@ -38,7 +33,6 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.Dialogs
             Telemetry.Instance.Native.TrackEvent("Connect.OK", new Dictionary<string, string>
             {
                 {"ProviderType", Model.ProviderType.ToString() },
-                {"UseCurrentUser", Model.UseCurrentUser.ToString() }
             });
 
             var client = SharePointProviderFactory.Create(Model.ProviderType);
@@ -48,22 +42,20 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.Dialogs
                 Model.IsConnecting = true;
                 StatusNotification.NotifyWithProgress("Connecting...");
 
-                var userName = !Model.UseCurrentUser ? Model.UserName : null;
-                var userPassword = !Model.UseCurrentUser ? Model.UserPassword: null;
-
-                if (await client.Connect(Model.SharePointWebUrl, userName, userPassword) != null)
+                if (await client.Connect(Model.SharePointWebUrl, Model.UserName, Model.UserPassword) != null)
                 {
                     Client = client;
                     Model.AddNewUrl(Model.SharePointWebUrl);
-                    Model.AddUserToHistory();
+                    Model.AddUserToHistory(Model.UserName);
                     Model.Save();
-                    DialogResult = true;
+
+                    DialogResult(this, client);
                 }
                 StatusNotification.Notify("Connected");
             }
             catch (Exception ex)
             {
-                ExceptionHandler.HandleConnection(ex);
+                Model.ErrorMessage = "Error: " + ExceptionHandler.HandleConnection(ex);
             }
 
             Model.IsConnecting = false;
@@ -72,7 +64,7 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.Dialogs
         private void ucCancelButton_Click(object sender, RoutedEventArgs e)
         {
             Telemetry.Instance.Native.TrackEvent("Connect.Cancel");
-            DialogResult = false;
+            DialogResult(this, null);
         }
 
         private void AdvanceOptionsButton_Click(object sender, RoutedEventArgs e)
@@ -89,7 +81,7 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.Dialogs
 
         private void PasswordBox_OnPasswordChanged(object sender, RoutedEventArgs e)
         {
-            Model.UserPassword = ((PasswordBox) sender).Password;
+            Model.UserPassword = ((PasswordBox)sender).Password;
         }
 
         private void ConnectWindow_OnClosing(object sender, CancelEventArgs e)